requirements.txt
asgiref==3.6.0
Django==4.2
django-bootstrap-form==3.4
django-cleanup==7.0.0
django-filter==23.2
django-spurl==0.6.8
packaging==23.0
Pillow==9.5.0
psycopg2==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
URLObject==2.4.3
